backend/backend_app/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# AI and Image Processing
deepface
opencv-python
pillow
numpy

# Optional AI packages (may require additional setup)
# tensorflow
# torch
# transformers


